Fixing renderSectionHeader type + useGetPendingTransactions
