+++ src/module/activity/queries/useGetPendingTransactions.ts
@@ -4,8 +4,14 @@ import { useQuery } from "react-query";
 import Queries from "../../../query/queries";
 import { TransactionType } from "module/sdk";
 import { TransactionRequestDto } from "module/activity/dto/dtos";
+import useFormatDate from "module/common/hook/useFormatDate";
 
-export default function (): QueryResult<TransactionRequestDto[]> {
+export interface ParsedPendingTransaction {
+    title: string;
+    data: TransactionRequestDto[];
+}
+
+export default function (): QueryResult<ParsedPendingTransaction[]> {
     const mockRequest: TransactionRequestDto = {
         transactionToken: "0",
         status: "pending",
@@ -36,12 +42,32 @@ export default function (): QueryResult<TransactionRequestDto[]> {
         mockRequest2,
     ];
 
+    const formatDate = useFormatDate();
+
+    const parsedPendingTransactions = (data: TransactionRequestDto[]): ParsedPendingTransaction[] => {
+        const timestampsMap: Record<string, TransactionRequestDto[]> = {};
+        data.map((transaction) => {
+            const { createdAt } = transaction;
+            const formattedDate = formatDate(createdAt);
+            if (!timestampsMap[formattedDate]) {
+                timestampsMap[formattedDate] = [transaction];
+            } else {
+                timestampsMap[formattedDate].push(transaction);
+            }
+        });
+        const timestamps = Object.keys(timestampsMap);
+        return timestamps.map((timestamp) => {
+            return { title: timestamp, data: timestampsMap[timestamp] };
+        });
+    };
+
     const { index: usedIndex, network, queryEnabled } = useServiceInstance();
     const getMockTransactionRequest = () => {
         return mockData;
     };
 
     return useQuery([Queries.SIGNER_APP_GET_PENDING_TRANSACTIONS, usedIndex, network], getMockTransactionRequest, {
+        select: (data) => parsedPendingTransactions(data),
         enabled: queryEnabled,
     });
 }
