+++ src/module/main/component/display/MainSectionList/MainSectionList.tsx
@@ -1,11 +1,16 @@
 import useCkbSync from "module/wallet/hook/useCkbSync";
 import { ListSection } from "module/main/component/display/MainSectionList/MainSectionList.styles";
 import { SectionList, SectionListProps, Typography } from "@peersyst/react-native-components";
+import { SectionListData } from "react-native";
 
-export type MainSectionListProps<T, D> = Omit<
+export type MainSectionListProps<T, D = any> = Omit<
     SectionListProps<T, D>,
     "stickySectionHeadersEnabled" | "contentContainerStyle" | "style" | "ItemSeparatorComponent"
 >;
+
+export interface RenderSectionHeaderProps<T> {
+    section: SectionListData<T, any>;
+}
 
 function MainSectionList<T, D>({ loading, onRefresh, ...rest }: MainSectionListProps<T, D>): JSX.Element {
     const { synchronizing, synchronize } = useCkbSync();
@@ -14,7 +19,7 @@ function MainSectionList<T, D>({ loading, onRefresh, ...rest }: MainSectionListP
         onRefresh?.();
     };
 
-    const renderSectionHeader = ({ section: { title } }: any): JSX.Element => {
+    const renderSectionHeader = ({ section: { title } }: RenderSectionHeaderProps<T>): JSX.Element => {
         return (
             <ListSection>
                 <Typography variant="body3Strong">{title}</Typography>
